new commit

+++ app/src/main/java/com/example/qrcodescanner/coding/DataClasses/ExtraPoint.kt
@@ -2,5 +2,5 @@ package com.example.qrcodescanner.coding.DataClasses
 
 data class ExtraPoint(
     val traineeId: String,
-    val point: Byte,
+    val point: Int,
 )
